# docs/admin/swagger.yaml
basePath: /
definitions:
...
  admin_app.ImageIdResponse:
    properties:
      id:
        type: string
    type: object
...
  common.ErrorResponse:
    properties:
      error:
        type: string
      msg:
        type: string
    type: object
host: localhost:8081
info:
  contact:
    email: email@email.com
    name: MatheusGomes
  description: This is the admin API for the Urchin app.
  license:
    name: MIT
  title: Urchin Admin API
  version: 1.0.0
paths:
  /images:
    post:
      consumes:
      - multipart/form-data
      description: Uploads an image file, saves it, and creates minified versions.
      parameters:
      - description: The image file to upload
        in: formData
        name: file
        required: true
        type: file
      - description: A brief description of the image
        in: formData
        name: excerpt
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin_app.ImageIdResponse'
        "400":
          description: Invalid input, file type, or size
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Server error while saving file
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Upload a new image
      tags:
      - images
  /images/{name}:
    delete:
      consumes:
      - application/json
      description: Deletes an image file from the server by its filename.
      parameters:
      - description: Image filename to delete
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin_app.ImageIdResponse'
        "400":
          description: Invalid or missing filename
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Delete an image
      tags:
      - images
...
schemes:
- http
swagger: "2.0"